Formated float arithmetic example
This an example on how the arithmetic result of float/double values
are affected based on the formating. This example will show that
formating can automatically ceil your result, which can affect
your outputs.

There are also ceil and floor function usage examples from the
math.h library, which will ceil or floor to the decimal point.