Different ways of defining for statements
You can have no expressions to multiple expressions in a for
statement. This example shows that you can have none to multiple
expressions for initialization and post-executions region for
for-statements.

You can also have constant values in while and for statements,
including assignments as well. Assignments will use the assigned
value as an outcome of the assignment expression. Also, please
remember that non-zero values are True, and zero value is False!