Short circuit example added
While using logical operators, depending on the logical
operator, the next expression may not be executed. For instance,
in an if-statement where there are two expression given as
condition connected with an AND operator, if the first expression
is false, the second expression does not get executed This is
because of the AND logic operator. If at leastone of the
expression is FALSE, no matter what the other expressions are,
the result will be FALSE, this is an example of short circuit.
Try executing the code, and see if the second expression will
be executed by checking the print result.