Examples of static modifier usage
You can use static for variable and for function declaration. If
a local variable is defined static, it behaves like global
variable, but can only be accesses where it is defined. It will
preserve progress.

On the other hand, if you define a function static, it can only be
accessed in the file where it is implemented. It cannot be called
from another function that is in another file.