sizeof on struct types
The size of the custom defined structs is determined by the members.
However, please keep in mind that the order of the members can
increase/decrease your type of struct. To keep the memory allocation
efficient with less redundancy we follow a naive approach by
defining the members smallest to largest. This is will not always
provide you the best solution, but still give you a good size.