Pre/post increment and decrement examples
Pre-increment/decrement first performs the increment or decrement,
and updates the value. Post on the other hand will do the increment
or the decrement after. However, if they are used standalone codes,
they are not different. The difference appears when you combine it
with other codes, such as; using it with printf, if, while, etc.