For statement definition with commas
For statement is based on main three sections. First is executed
once and can include multiple initialization sepearated with comma.
Second section is the condition that will check at the begining and
after each end of loop. The third section is the statements that
are executed at the end, after all the codes within the for loop
block is executed. You can add multiple instructions on the third
section by separating them with commas.